fix: fixed custom search hook in week 3
