+++ src/components/Home/Aditya/Week3/components/searchUser/SearchUser.tsx
@@ -1,30 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { SearchUserWrapper } from './searchUserStyles';
+import statusChoices from '../../../../Keshav/Week3/utils/statusOptions';
 import useSearch from '../../hooks/useSearch';
-import { statusChoices } from '../../utils/constants';
-import { User } from '../../interfaces/interfaces';
+import { searchResult } from '../../interfaces/interfaces';
 
 const SearchUser: React.FC = () => {
-  const [search, setSearch] = useState<string>('');
-  const results: User[] = useSearch(search);
-
-  const handleSearch = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    setSearch(e.target.value);
-  };
-
-  const handleStatusChange = (e: React.ChangeEvent<HTMLSelectElement>): void => {
-    const { name, value } = e.target;
-    const users = JSON.parse(localStorage.getItem('users') ?? '[]');
-    const updatedUsers = users.map((user: User) =>
-      user.name === name
-        ? {
-            name: name,
-            status: value
-          }
-        : user
-    );
-    localStorage.setItem('users', JSON.stringify(updatedUsers));
-  };
+  const { search, setSearch, data, handleStatusChange }: searchResult = useSearch();
 
   return (
     <SearchUserWrapper>
@@ -33,9 +14,15 @@ const SearchUser: React.FC = () => {
         <label htmlFor='search' className='fontMedium'>
           Search:
         </label>
-        <input type='text' className='userDetail' name='search' value={search} onChange={handleSearch} />
-        <p>{!results.length ? 'No results' : 'Results:'}</p>
-        {results.map((user) => (
+        <input
+          type='text'
+          className='userDetail'
+          name='search'
+          value={search}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => setSearch(e.target.value)}
+        />
+        <p>{!data.length ? 'No results' : 'Results:'}</p>
+        {data.map((user) => (
           <div className='results' key={user.name}>
             <span>{user.name}</span>
             <select id='userStatus' className='status' name={user.name} value={user.status} onChange={handleStatusChange}>
