+++ src/components/Home/Aditya/Week4/pages/Week3/interfaces/interfaces.ts
@@ -6,3 +6,10 @@ export interface User {
   name: string;
   status: string;
 }
+
+export interface searchResult {
+  search: string;
+  setSearch: React.Dispatch<React.SetStateAction<string>>;
+  data: User[];
+  handleStatusChange: (e: React.ChangeEvent<HTMLSelectElement>) => void;
+}
